# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.CalculationGroupResponse:
    description: Информация о группе в составе заявки
    properties:
      cavi_index:
        example: 7.5
        type: number
      group:
        $ref: '#/definitions/handler.GroupResponse'
      group_id:
        example: 1
        type: integer
    type: object
  handler.CalculationResponse:
    description: Полная информация о заявке на расчёт CAVI
    properties:
      calculation_groups:
        items:
          $ref: '#/definitions/handler.CalculationGroupResponse'
        type: array
      creator_login:
        example: user1
        type: string
      date_completed:
        example: "2025-01-15T12:00:00Z"
        type: string
      date_created:
        example: "2025-01-15T10:30:00Z"
        type: string
      date_formed:
        example: "2025-01-15T11:00:00Z"
        type: string
      diastolic_pressure:
        example: 80
        type: integer
      groups_count:
        example: 3
        type: integer
      id:
        example: 1
        type: integer
      moderator_login:
        example: moderator1
        type: string
      pulse_wave_velocity:
        example: 8.5
        type: number
      status:
        example: draft
        type: string
      systolic_pressure:
        example: 120
        type: integer
    type: object
  handler.CalculationUpdateRequest:
    description: Данные для обновления полей заявки (давление, скорость пульсовой
      волны)
    properties:
      diastolic_pressure:
        example: 80
        type: integer
      pulse_wave_velocity:
        example: 8.5
        type: number
      systolic_pressure:
        example: 120
        type: integer
    type: object
  handler.CartIconResponse:
    description: Информация о черновике заявки пользователя
    properties:
      calculation_id:
        example: 1
        type: integer
      items:
        example: 3
        type: integer
    type: object
...
  handler.ErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      message:
        example: invalid input
        type: string
      status:
        example: fail
        type: string
    type: object
...
  handler.GroupResponse:
    description: Информация о группе пациентов
    properties:
      age_group:
        example: young
        type: string
      description:
        example: Эластичные сосуды с низким уровнем жесткости
        type: string
      disease_type:
        example: diabetes
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: http://localhost:9000/cavi-images/diagrams/1.jpg
        type: string
      is_deleted:
        example: false
        type: boolean
      is_selected:
        example: false
        type: boolean
      name:
        example: Молодые пациенты (до 35 лет)
        type: string
    type: object
...
  handler.ModerateRequest:
    description: Действие модератора над заявкой
    properties:
      action:
        enum:
        - complete
        - reject
        example: complete
        type: string
    required:
    - action
    type: object
...
  handler.SuccessResponse:
    description: Стандартный успешный ответ
    properties:
      status:
        example: ok
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cavi.local
    name: API Support
  description: |-
    API для расчёта индекса CAVI (Cardio-Ankle Vascular Index).
    Система позволяет управлять группами пациентов, создавать заявки на расчёт,
    добавлять группы в заявки и проводить модерацию.
  title: CAVI Calculator API
  version: "1.0"
paths:
  /cavi-calculations:
    get:
      description: Возвращает список заявок (кроме удалённых и черновиков) с фильтрацией
        по статусу и диапазону дат формирования.
      parameters:
      - description: Фильтр по статусу
        enum:
        - formed
        - completed
        - rejected
        in: query
        name: status
        type: string
      - description: Дата формирования от (YYYY-MM-DD)
        example: '"2025-01-01"'
        in: query
        name: date_from
        type: string
      - description: Дата формирования до (YYYY-MM-DD)
        example: '"2025-12-31"'
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            items:
              $ref: '#/definitions/handler.CalculationResponse'
            type: array
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список заявок
      tags:
      - calculations
  /cavi-calculations/{id}:
    delete:
      description: Мягкое удаление заявки-черновика. Доступно только создателю заявки.
        Удаляет связи с группами.
      parameters:
      - description: ID заявки
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заявка удалена
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Неверный ID или заявка не является черновиком
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить заявку
      tags:
      - calculations
    get:
      description: Возвращает детальную информацию о заявке, включая все группы в
        её составе.
      parameters:
      - description: ID заявки
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о заявке
          schema:
            $ref: '#/definitions/handler.CalculationResponse'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить заявку
      tags:
      - calculations
    put:
      consumes:
      - application/json
      description: 'Обновляет поля заявки: систолическое/диастолическое давление,
        скорость пульсовой волны.'
      parameters:
      - description: ID заявки
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CalculationUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Заявка обновлена
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить заявку
      tags:
      - calculations
  /cavi-calculations/{id}/moderate:
    put:
      consumes:
      - application/json
      description: Завершает или отклоняет сформированную заявку. Доступно только
        модераторам. При завершении устанавливается количество групп.
      parameters:
      - description: ID заявки
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Действие модератора
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ModerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Заявка обработана
          schema:
            $ref: '#/definitions/handler.CalculationResponse'
        "400":
          description: Неверные данные или статус заявки
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Требуется роль модератора
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Модерировать заявку
      tags:
      - calculations
  /cavi-calculations/draft:
    get:
      description: Возвращает ID черновика заявки текущего пользователя и количество
        групп в нём.
      produces:
      - application/json
      responses:
        "200":
          description: Информация о корзине
          schema:
            $ref: '#/definitions/handler.CartIconResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Корзина (черновик)
      tags:
      - calculations
  /cavi-calculations/draft/form:
    put:
      description: Переводит черновик в статус "сформирована". Проверяет наличие групп
        в заявке. Устанавливает дату формирования.
      produces:
      - application/json
      responses:
        "200":
          description: Заявка сформирована
          schema:
            $ref: '#/definitions/handler.CalculationResponse'
        "400":
          description: Нет групп в заявке
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Черновик не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Сформировать заявку
      tags:
      - calculations
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"